updates in typescript code
