+++ src/components/SectionContent.vue
@@ -39,15 +39,7 @@
 
 <script lang="ts">
 
-interface Data {
-  title_h2: string;
-  description: string;
-  call_to_action: Action;
-  image: Image;
-}
-
-import Action from '../typescript/action';
-import Image from '../typescript/action';
+import Data from '../typescript/data';
 import { defineComponent, PropType } from 'vue';
 
 export default defineComponent({
