+++ src/components/SectionWithBuckets.vue
@@ -22,19 +22,7 @@
 
 <script lang="ts">
 
-interface Buckets {
-  icon: Image;
-}
-
-interface Data {
-  title_h2: string;
-  description: string;
-  buckets: Buckets;
-  call_to_action: Action;
-}
-
-import Image from '../typescript/action';
-import Action from '../typescript/action';
+import Data from '../typescript/data';
 import { defineComponent, PropType } from 'vue';
 
 export default defineComponent({
